requirements.txt
azure_storage
fabric_cicd
python-dotenv
ms-fabric-cli